requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
groq==0.13.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.3
openpyxl==3.1.5
